Revert "Doc upd"

+++ src/include/lock/spinlock.h
@@ -1,9 +1,7 @@
 #pragma once
 #define _SPINLOCK_H_
 
-#include <stdint.h>
-
-typedef size_t spinlock_t;
+typedef unsigned int spinlock_t;
 
 void spinlock_acquire(spinlock_t *);
 void spinlock_release(spinlock_t *);
